+++ NEWS.md
@@ -1,3 +1,8 @@
+# ghqc.app 0.4.17
+
+- allows binary files to be assigned in `ghqc_assign_app()`
+- fixes bug in which file names with spaces aren't recognized
+
 # ghqc.app 0.4.16
 
 - fixes bug in which tables don't hold position in QC Records